possibly working parser

+++ src/tree_serialize.rs
@@ -1,4 +1,4 @@
-use super::message_generated::asted::interface::tree::{Location, Node, NodeArgs};
+use super::message_generated::asted::interface::{Location, Node, NodeArgs};
 use flatbuffers::{self, WIPOffset};
 use tree_sitter;
 
@@ -9,6 +9,8 @@ pub fn serialize(text: &[u16], tree: &tree_sitter::Tree) -> Vec<u8> {
 	let root_node = build_node(text, &mut builder, tree.root_node());
 
 	builder.finish_size_prefixed(root_node, None);
+	// TODO(sauyon): to_vec is a copy, need to bubble up the builder to the actual handler function
+	//               since the builder doesn't have any functions to take ownership of the buffer
 	builder.finished_data().to_vec()
 }
 
@@ -18,7 +20,7 @@ fn build_node<'a>(
 	node: tree_sitter::Node<'a>,
 ) -> WIPOffset<Node<'a>> {
 	let kind = builder.create_string(node.kind());
-	let location = Location::new(node.start_byte() as u64, node.end_byte() as u64);
+	let location = Location::new(node.start_byte() as u32, node.end_byte() as u32);
 	let child_vec = node.children(&mut node.walk()).map(|child| build_node(text, builder, child)).collect::<Vec<_>>();
 	let children = builder.create_vector(&child_vec);
